Simplified reading and writing of statistical models
